+++ CHANGES.md
@@ -116,6 +116,8 @@ through the attributes DlCtrlSymbols and UlCtrlSymbols.
 
 * GetN2Delay moved to MmWaveEnbPhy.
 
+* Removed attribute L1L2CtrlLatency, and fixed it to 2 in MmWavePhy.
+
 ### Changed behavior:
 
 * K0, K1, K2 Delays are removed from the phy-mac common, instead they are